+++ src/types/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './auth.types.js';
+export * from './response.types.js';